Support clients using CMake
This declares the version of the library to be 1.0 and enables proper
find_package usage by consuming projects. Versioning is required to not
produce broken CMake packages that have no version config scripts.

The package name of the library is "plutovg" and the exported library
target's name is "plutovg::plutovg".